requirements.txt
mysql-connector-python==8.0.32
requests==2.28.2